Initial project commitment
After a fun-filled evening, we have a basic assembler written entirely
as a CoffeeScript application for a web server. The project is
completely 8-bit, and even the addressing can't exceed 8 bits. ;-)

Hooked up addresses #81, #82, #83 and #84 to "lights" that have an RGB
value associated with the first three pins. Yeah, I'll have to work on a
tutorial or this won't really make sense.