Better sections and sub-sections
Each section, while being remember and re-displayed when re-logging in,
it also downloaded dynamically from another web page file (also removing
the H1 and other non-helpful tags that do not belong in an embedded
page).

Links initially bring up a dialog box allowing us to hold the attention
of the reader and not send them elsewhere, but also bringing them back
to the lesson plan.